Quick & dirty attempt to preserve the original config files

+++ install.sh
@@ -10,21 +10,21 @@ ln -sfn $DOTFILES/bashrc ~/.bashrc
 touch ~/.bashrc.local
 
 # vim
-mv ~/.vimrc ~/.vimrc.old 2> /dev/null
-mv ~/.vim ~/.vim.old 2> /dev/null
-ln -s $DOTFILES/vimrc ~/.vimrc
-ln -s $DOTFILES/vim ~/.vim
+test -f ~/.vimrc && mv ~/.vimrc ~/.vimrc.old 2> /dev/null
+test -d ~/.vim && mv ~/.vim ~/.vim.old 2> /dev/null
+ln -sfn $DOTFILES/vimrc ~/.vimrc
+ln -sfn $DOTFILES/vim ~/.vim
 
 # screen
-mv ~/.screenrc ~/.screenrc.old 2> /dev/null
-ln -s $DOTFILES/screenrc ~/.screenrc
+test -f ~/.screenrc && mv ~/.screenrc ~/.screenrc.old 2> /dev/null
+ln -sfn $DOTFILES/screenrc ~/.screenrc
 
 # tmux
-#mv ~/.tmux.conf ~/.tmux.conf.old 2> /dev/null
+#test -f ~/.tmux.conf && mv ~/.tmux.conf ~/.tmux.conf.old 2> /dev/null
 #ln -s $DOTFILES/tmux.conf ~/.tmux.conf
 
 # dircolors
-#mv ~/.dir_colors ~/.dir_colors.old 2> /dev/null
+#test -f ~/.dir_colors && mv ~/.dir_colors ~/.dir_colors.old 2> /dev/null
 #ln -s $DOTFILES/dircolors/dircolors.ansi-dark ~/.dir_colors
 
 #git submodule update --init
